#760: move task list files to projectviewcontainer
